+++ src/components/returnButton/ReturnButton.style.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const ChangesButton = styled.button`
+    width: 300px;
+    font-size: 1.6rem;
+    padding: 10px;
+    border: 1px solid black;
+    border-radius: 5px;
+    cursor: pointer;
+`;
+
+export { ChangesButton };
