+++ src/routes/vendingMachine/VendingMachine.style.js
@@ -14,13 +14,4 @@ const UserInterfaceContainer = styled.div`
     gap: 20px;
 `;
 
-const ChangesButton = styled.button`
-    width: 300px;
-    font-size: 1.6rem;
-    padding: 10px;
-    border: 1px solid black;
-    border-radius: 5px;
-    cursor: pointer;
-`;
-
-export { VendingMachineContainer, UserInterfaceContainer, ChangesButton };
+export { VendingMachineContainer, UserInterfaceContainer };
